+++ src/main/groovy/ru/vyarus/gradle/plugin/animalsniffer/util/AndroidClassesCollector.groovy
@@ -3,11 +3,10 @@ package ru.vyarus.gradle.plugin.animalsniffer.util
 import groovy.transform.CompileStatic
 import org.gradle.api.DefaultTask
 import org.gradle.api.file.Directory
-import org.gradle.api.file.DirectoryProperty
 import org.gradle.api.file.RegularFile
 import org.gradle.api.provider.ListProperty
+import org.gradle.api.provider.SetProperty
 import org.gradle.api.tasks.InputFiles
-import org.gradle.api.tasks.OutputDirectory
 import org.gradle.api.tasks.TaskAction
 
 /**
@@ -19,25 +18,22 @@ import org.gradle.api.tasks.TaskAction
 @CompileStatic
 abstract class AndroidClassesCollector extends DefaultTask {
 
+    static String computeTaskName(String prefix) {
+        prefix + 'AnimalsnifferClassesCollector'
+    }
+
     @InputFiles
     abstract ListProperty<RegularFile> getJarFiles()
 
     @InputFiles
     abstract ListProperty<Directory> getClassesDirs()
 
-    @OutputDirectory
-    abstract DirectoryProperty getOutputDirectory()
-
-    @SuppressWarnings('AbstractClassWithPublicConstructor')
-    AndroidClassesCollector() {
-        outputDirectory.value(project.layout.buildDirectory.dir('intermediates/animal_sniffer/' + name))
-    }
+    @InputFiles
+    abstract SetProperty<File> getMultiplatformSourceDirs()
 
     @TaskAction
+    @SuppressWarnings('EmptyMethodInAbstractClass')
     void execute() {
-        project.sync {
-            it.from(classesDirs)
-            it.into(outputDirectory)
-        }
+        // no body
     }
 }