+++ src/test/groovy/ru/vyarus/gradle/plugin/animalsniffer/android/multiplatform/KotlinMultiplatformMultipleTargetsAndroidKitTest.groovy
@@ -7,17 +7,19 @@ import spock.lang.IgnoreIf
 
 /**
  * @author Vyacheslav Rusakov
- * @since 23.11.2024
+ * @since 11.12.2024
  */
 @IgnoreIf({ !jvm.java17Compatible })
-class KotlinMultiplatformMultipleTargetsKitTest extends AbstractAndroidKitTest {
+class KotlinMultiplatformMultipleTargetsAndroidKitTest extends AbstractAndroidKitTest {
 
-    def "Check multiple declared platforms (no android)"() {
+    def "Check multiple declared platforms (with android)"() {
         setup:
         build """
             // this is necessary to avoid the plugins to be loaded multiple times
             // in each subproject's classloader            
             plugins {
+                id 'com.android.application' version '8.4.0' apply false
+                id 'com.android.library' version '8.4.0' apply false
                 id 'org.jetbrains.compose' version '1.7.0' apply false
                 id 'org.jetbrains.kotlin.plugin.compose' version '2.0.21' apply false
                 id 'org.jetbrains.kotlin.jvm' version '2.0.21' apply false
@@ -26,28 +28,53 @@ class KotlinMultiplatformMultipleTargetsKitTest extends AbstractAndroidKitTest {
             }            
         """
 
-        file('shared/build.gradle') << """
+        file('shared/build.gradle') << """      
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
+
 plugins {
     id 'org.jetbrains.kotlin.multiplatform'
+    id 'com.android.library'
     id 'ru.vyarus.animalsniffer'
 }
 
-animalsniffer {
-    ignoreFailures = true
-}
-
 kotlin {
+    androidTarget {
+        compilerOptions {
+            jvmTarget = JvmTarget.JVM_17
+        }
+    }
+
     jvm()
     
     sourceSets {
-        commonMain.dependencies {
-            implementation 'org.slf4j:slf4j-api:1.7.25'          
+        commonMain.dependencies {                     
+            implementation 'org.slf4j:slf4j-api:1.7.25'
         }
-    }
+    }  
 }
 
 dependencies {
     signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
+}
+
+android {
+    namespace = "org.example.project.shared"
+    compileSdk = 34
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+    defaultConfig {
+        minSdk = 24
+    }
+    lint {
+        checkReleaseBuilds false
+        abortOnError false
+    }
+}
+
+animalsniffer {
+    ignoreFailures = true
 }
 """
         fileFromClasspath('shared/src/commonMain/kotlin/invalid/Sample.kt', '/ru/vyarus/gradle/plugin/animalsniffer/kotlin/invalid/Sample.kt')
@@ -91,9 +118,11 @@ dependencies {
 
         file('composeApp/build.gradle') << """    
 import org.jetbrains.compose.desktop.application.dsl.TargetFormat
+import org.jetbrains.kotlin.gradle.dsl.JvmTarget
 
 plugins {
     id 'org.jetbrains.kotlin.multiplatform'
+    id 'com.android.application'
     id 'org.jetbrains.compose'
     id 'org.jetbrains.kotlin.plugin.compose'
     id 'ru.vyarus.animalsniffer'
@@ -104,9 +133,20 @@ animalsniffer {
 }
 
 kotlin {
+    androidTarget {
+        compilerOptions {
+            jvmTarget == JvmTarget.JVM_17
+        }
+    }
+
     jvm("desktop")
     
     sourceSets {
+    
+        androidMain.dependencies {
+            implementation(compose.preview)
+            implementation 'androidx.activity:activity-compose:1.9.3'
+        }
         
         commonMain.dependencies {
             implementation(compose.runtime)
@@ -117,16 +157,49 @@ kotlin {
             implementation(compose.components.uiToolingPreview)
             implementation 'org.jetbrains.androidx.lifecycle:lifecycle-viewmodel:2.8.3'
             implementation 'org.jetbrains.androidx.lifecycle:lifecycle-runtime-compose:2.8.3'
+            implementation 'org.slf4j:slf4j-api:1.7.25'
             implementation project(':shared')
-            implementation 'org.slf4j:slf4j-api:1.7.25' 
         }
         desktopMain.dependencies {
             implementation(compose.desktop.currentOs)
             implementation 'org.jetbrains.kotlinx:kotlinx-coroutines-swing:1.9.0'
+            implementation 'org.slf4j:slf4j-api:1.7.25'
         }
     }
 }
 
+android {
+    namespace = "org.example.project"
+    compileSdk = 34
+
+    defaultConfig {
+        applicationId = "org.example.project"
+        minSdk = 24
+        targetSdk = 34
+        versionCode = 1
+        versionName = "1.0"
+    }
+    buildTypes {
+        getByName("release") {
+            minifyEnabled = false
+        }
+    }
+    compileOptions {
+        sourceCompatibility = JavaVersion.VERSION_17
+        targetCompatibility = JavaVersion.VERSION_17
+    }
+    lint {
+        quiet true
+        checkReleaseBuilds false
+        abortOnError false
+    }
+}
+
+dependencies {
+    signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
+
+    debugImplementation(compose.uiTooling)
+}
 
 compose.desktop {
     application {
@@ -139,18 +212,22 @@ compose.desktop {
         }
     }
 }
-
-dependencies {
-    signature 'org.codehaus.mojo.signature:java16-sun:1.0@signature'
-}
 """
         fileFromClasspath('composeApp/src/commonMain/kotlin/invalid/Sample.kt', '/ru/vyarus/gradle/plugin/animalsniffer/kotlin/invalid/Sample.kt')
         fileFromClasspath('composeApp/src/desktopMain/kotlin/invalid/Sample2.kt', '/ru/vyarus/gradle/plugin/animalsniffer/kotlin/invalid/Sample2.kt')
+        fileFromClasspath('composeApp/src/androidMain/kotlin/invalid/Sample3.kt', '/ru/vyarus/gradle/plugin/animalsniffer/kotlin/invalid/Sample3.kt')
 
+        generateManifest('composeApp/src/androidMain')
         file("settings.gradle") << """            
             include ':composeApp', ':server', ':shared'
             """
+        file("gradle.properties") << """ 
+android.nonTransitiveRClass=true
+android.useAndroidX=true
+io.ktor.development=true
+"""
 
+//        fileFromClasspath('src/androidMain/kotlin/invalid/Sample.kt', '/ru/vyarus/gradle/plugin/animalsniffer/kotlin/invalid/Sample.kt')
 //        debug()
 
         when: "run shared task"
@@ -158,9 +235,9 @@ dependencies {
 
         then: "task successful"
         result.task(':shared:check').outcome == TaskOutcome.SUCCESS
-        result.task(':shared:animalsnifferJvmMain').outcome == TaskOutcome.SUCCESS
+        result.task(':shared:animalsnifferDebug').outcome == TaskOutcome.SUCCESS
 
-        then: "violations detected"
+        then: "validate shared report"
         result.output.contains("4 AnimalSniffer violations were found in 2 files")
         result.output.replaceAll('\r', '').contains(
                 """[Undefined reference] invalid.(Sample.kt:11)
@@ -184,7 +261,7 @@ dependencies {
         result.task(':server:check').outcome == TaskOutcome.SUCCESS
         result.task(':server:animalsnifferMain').outcome == TaskOutcome.SUCCESS
 
-        then: "violations detected"
+        then: "validate server report"
         result.output.contains("2 AnimalSniffer violations were found in 1 files")
         result.output.replaceAll('\r', '').contains(
                 """[Undefined reference] invalid.(Sample.kt:11)
@@ -195,6 +272,7 @@ dependencies {
 """)
 
 
+
         when: "run composeApp task"
         result = run(':composeApp:check')
 
@@ -202,7 +280,7 @@ dependencies {
         result.task(':composeApp:check').outcome == TaskOutcome.SUCCESS
         result.task(':composeApp:animalsnifferDesktopMain').outcome == TaskOutcome.SUCCESS
 
-        then: "violations detected"
+        then: "validate compose report"
         result.output.contains("4 AnimalSniffer violations were found in 2 files")
         result.output.replaceAll('\r', '').contains(
                 """[Undefined reference] invalid.(Sample.kt:11)
@@ -218,5 +296,4 @@ dependencies {
   >> Iterable java.nio.file.FileSystem.getFileStores()
 """)
     }
-
 }